Rename maven artifact and update README
