+++ pom.xml
@@ -2,8 +2,25 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.vanillaforums</groupId>
-    <artifactId>vanilla</artifactId>
-    <version>2.1</version>
+    <artifactId>jsconnect</artifactId>
+    <version>3.0</version>
+    <packaging>jar</packaging>
+
+    <name>jsConnect</name>
+    <description>Client libraries to use the jsConnect SSO protocol with Vanilla Forums.</description>
+
+    <licenses>
+        <license>
+            <name>MIT</name>
+            <url>https://github.com/vanilla/jsConnectJava/blob/master/LICENSE.md</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+
+    <scm>
+        <url>https://github.com/vanilla/jsConnectJava</url>
+    </scm>
+
     <build>
         <plugins>
             <plugin>
@@ -34,9 +51,7 @@
             </plugin>
         </plugins>
     </build>
-    <packaging>jar</packaging>
-    <name>vanillaforums</name>
-    <url>http://maven.apache.org</url>
+
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
