+++ src/main/java/com/vanillaforums/jsconnect/FieldNotFoundException.java
@@ -1,4 +1,4 @@
-package com.vanillaforums.vanilla;
+package com.vanillaforums.jsconnect;
 
 public class FieldNotFoundException extends JsConnectException {
     public FieldNotFoundException(String field, String collection) {
