+++ src/main/java/com/vanillaforums/jsconnect/InvalidValueException.java
@@ -1,4 +1,4 @@
-package com.vanillaforums.vanilla;
+package com.vanillaforums.jsconnect;
 
 public class InvalidValueException extends JsConnectException {
     public InvalidValueException(String message) {
