+++ src/main/java/com/vanillaforums/jsconnect/JsConnectException.java
@@ -1,4 +1,4 @@
-package com.vanillaforums.vanilla;
+package com.vanillaforums.jsconnect;
 
 public class JsConnectException extends Exception {
     public JsConnectException(String message) {
