+++ src/test/java/com/vanillaforums/jsconnect/JsConnectServlet.java
@@ -1,4 +1,4 @@
-package com.vanillaforums.vanilla;
+package com.vanillaforums.jsconnect;
 
 import org.apache.commons.text.StringEscapeUtils;
