+++ src/test/java/com/vanillaforums/jsconnect/JsConnectServletV3.java
@@ -1,14 +1,12 @@
-package com.vanillaforums.vanilla;
+package com.vanillaforums.jsconnect;
 
 import org.apache.commons.text.StringEscapeUtils;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.net.URI;
-import java.util.HashMap;
 
 public class JsConnectServletV3 extends HttpServlet {
     private String clientID, secret;
