+++ src/test/java/com/vanillaforums/jsconnect/SSOStringTest.java
@@ -1,4 +1,4 @@
-package com.vanillaforums.vanilla;
+package com.vanillaforums.jsconnect;
 
 import org.junit.jupiter.api.Test;
